// api/src/unidade-atendimento/dto/resposta-unidade-atendimento.dto.ts
/** Detalhe — BigInt como string no JSON. */
export class RespostaUnidadeAtendimentoDto {
  id: string;
  id_tenacidade: string;
  idAlmoxarifadoPadrao: string | null;
  idMunicipioNotaFiscal: string | null;
  razaoSocial: string | null;
  nomeFantasia: string | null;
  cnpj: string | null;
  inscricaoMunicipal: string | null;
  naturezaOperacao: number | null;
  regimeTributacao: number | null;
  serieNotaFiscal: string | null;
  cnes: string | null;
  sigla: string | null;
  tipoUnidade: string | null;
  cep: string | null;
  tipoLogradouro: string | null;
  logradouro: string | null;
  bairro: string | null;
  numero: string | null;
  complemento: string | null;
  cidade: string | null;
  estado: string | null;
  latitude: number | null;
  longitude: number | null;
  telefone: string | null;
  celular: string | null;
  amostraExternaInicio: string | null;
  amostraExternaFim: string | null;
  amostraExternaAtual: string | null;
  senhaInternet: string | null;
  regraAgrupamentoAmostra: number | null;
  nomeArquivoLogotipo: string | null;
  nomeReferenciaLogotipo: string | null;
  enviarAmostra: string | null;
  enviarSms: string | null;
  procedenciaAtendimentoObrigatorio: string | null;
  descricaoServicoNotaFiscal: string | null;
  pertencerSimplesNacional: string | null;
  qtdDiasValidadeOrcamento: number | null;
  alterarLaboratorioApoio: string | null;
  listaProcedenciaPermitida: string | null;
  listaConvenioPermitido: string | null;
  indicacaoOrcamento: string | null;
  ativo: string | null;
  codigoExterno: string | null;
  id_usuario_auditoria: string | null;
  endereco_ip_auditoria: string | null;
  nome_aplicacao_auditoria: string | null;
}